requirements.txt
Django==4.1.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
